Reorganise to Api and Client, remove suppression time
