+++ src/Seq.App.Opsgenie/Api/OpsGenieResponse.cs
@@ -1,6 +1,6 @@
 ﻿// ReSharper disable UnusedMember.Global
 
-namespace Seq.App.Opsgenie.Classes
+namespace Seq.App.Opsgenie.Api
 {
     public class OpsGenieResponse
     {
