+++ src/Seq.App.Opsgenie/Api/OpsGenieResult.cs
@@ -5,7 +5,7 @@ using System.Net.Http;
 
 // ReSharper disable NotAccessedField.Global
 
-namespace Seq.App.Opsgenie.Classes
+namespace Seq.App.Opsgenie.Api
 {
     public class OpsGenieResult
     {
