+++ src/Seq.App.Opsgenie/Api/ResponderType.cs
@@ -1,7 +1,7 @@
 ﻿// Used to maintain the (JSON) OpsGenie API contract.
 // ReSharper disable UnusedMember.Global
 
-namespace Seq.App.Opsgenie.Classes
+namespace Seq.App.Opsgenie.Api
 {
     public enum ResponderType
     {
