+++ src/Seq.App.Opsgenie/OpsgenieApp.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Seq.App.Opsgenie.Api;
 using Seq.App.Opsgenie.Classes;
-using Seq.App.Opsgenie.Interfaces;
+using Seq.App.Opsgenie.Client;
 using Seq.Apps;
 using Seq.Apps.LogEvents;
 
@@ -32,7 +33,6 @@ namespace Seq.App.Opsgenie
         Priority _priority = Priority.P3, _defaultPriority = Priority.P3;
         string _priorityProperty = "@Level";
         string _responderProperty;
-        TimeSpan _suppressionTime;
         string[] _tags;
 
         // Permits substitution for testing.
@@ -43,14 +43,7 @@ namespace Seq.App.Opsgenie
             HelpText = "The Opsgenie API key to use.",
             InputType = SettingInputType.Password)]
         public string ApiKey { get; set; }
-
-        [SeqAppSetting(
-            DisplayName = "Suppression interval (seconds)",
-            HelpText =
-                "If an alert has been raised, further alerts will be suppressed for this time. Default 60, Minimum 0.",
-            InputType = SettingInputType.Integer)]
-        public int SuppressionTime { get; set; } = 60;
-
+        
         [SeqAppSetting(
             IsOptional = true,
             DisplayName = "Alert message",
@@ -142,8 +135,6 @@ namespace Seq.App.Opsgenie
         public async Task OnAsync(Event<LogEventData> evt)
         {
             if (evt == null) throw new ArgumentNullException(nameof(evt));
-            var suppressDiff = DateTime.UtcNow - _lastAlert;
-            if (suppressDiff.TotalSeconds < _suppressionTime.TotalSeconds) return;
 
             var message = _generateMessage.Render(evt);
             var description = _generateDescription.Render(evt);
@@ -217,12 +208,7 @@ namespace Seq.App.Opsgenie
                 !string.IsNullOrWhiteSpace(AlertDescription)
                     ? AlertDescription
                     : $"Generated by Seq running at {Host.BaseUri}.");
-
-            if (SuppressionTime < 0)
-                SuppressionTime = 0;
-            _suppressionTime = TimeSpan.FromSeconds(SuppressionTime);
-            _lastAlert = DateTime.UtcNow - _suppressionTime;
-
+            
             if (!string.IsNullOrEmpty(PriorityProperty))
                 _priorityProperty = PriorityProperty;
