+++ src/Seq.App.Opsgenie/Template/HandlebarsTemplate.cs
