+++ test/Seq.App.Opsgenie.Tests/OpsgenieAppTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Seq.App.Opsgenie.Api;
 using Seq.App.Opsgenie.Classes;
 using Seq.App.Opsgenie.Tests.Support;
 using Seq.Apps;
